requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
ccxt==4.4.89
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
Flask==3.1.1
flask-cors==6.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
multidict==6.5.0
numpy
pandas==2.3.0
propcache==0.3.2
pycares==4.9.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
requests==2.32.4
scikit-learn==1.7.0
scipy==1.15.3
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.41
threadpoolctl==3.6.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xgboost==3.0.2
yarl==1.20.1
python-telegram-bot==20.7
python-dotenv

